# docs/api/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "CandidatePage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateSummaryOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CandidatePage",
        "type": "object"
      },
      "CandidateResolutionIn": {
        "properties": {
          "decision": {
            "enum": [
              "checked",
              "dismiss"
            ],
            "title": "Decision",
            "type": "string"
          },
          "reason": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "pattern": "\\S",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "required": [
          "decision"
        ],
        "title": "CandidateResolutionIn",
        "type": "object"
      },
...
      "CandidateReviewItemPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CandidateReviewItemOut"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "CandidateReviewItemPage",
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "description": "Stable, for your code",
                "type": "string"
              },
              "details": {
                "description": "Depends on code",
                "type": "object"
              },
              "message": {
                "description": "For people; may change",
                "type": "string"
              },
              "request_id": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "JoinOut": {
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "joined_at": {
            "title": "Joined At",
            "type": "string"
          },
          "joined_by": {
            "title": "Joined By",
            "type": "string"
          },
          "joined_candidate_id": {
            "title": "Joined Candidate Id",
            "type": "string"
          },
          "primary_candidate_id": {
            "title": "Primary Candidate Id",
            "type": "string"
          },
          "reason": {
            "title": "Reason",
            "type": "string"
          },
          "undone_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Undone At"
          },
          "undone_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Undone By"
          },
          "undone_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Undone Reason"
          }
        },
        "required": [
          "id",
          "primary_candidate_id",
          "joined_candidate_id",
          "reason",
          "joined_by",
          "joined_at",
          "undone_at",
          "undone_by",
          "undone_reason"
        ],
        "title": "JoinOut",
        "type": "object"
      },
...
      "UndoJoinIn": {
        "properties": {
          "reason": {
            "maxLength": 500,
            "minLength": 1,
            "pattern": "\\S",
            "title": "Reason",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "UndoJoinIn",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "A sign-in token from the company identity provider, or from /dev/token in dev.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "Talent Platform",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/candidate-joins/{join_id}/undo": {
      "post": {
        "description": "Undoes a join, with a reason. The join itself is kept, marked undone by you.",
        "operationId": "undo_join_v1_candidate_joins__join_id__undo_post",
        "parameters": [
          {
            "in": "path",
            "name": "join_id",
            "required": true,
            "schema": {
              "title": "Join Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UndoJoinIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Undo Join",
        "tags": [
          "review queue",
          "candidates"
        ]
      }
    },
    "/v1/candidate-review-items": {
      "get": {
        "description": "Items about candidates in your scope, newest first. Open items by default.",
        "operationId": "list_candidate_review_items_v1_candidate_review_items_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "default": "open",
              "enum": [
                "open",
                "resolved"
              ],
              "title": "Status",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "flagged_document",
                    "unverified_candidate",
                    "possible_duplicate",
                    "borderline_score",
                    "ai_assessment"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "candidate_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReviewItemPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Candidate Review Items",
        "tags": [
          "review queue"
        ]
      }
    },
    "/v1/candidate-review-items/{review_item_id}": {
      "get": {
        "operationId": "get_candidate_review_item_v1_candidate_review_items__review_item_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "review_item_id",
            "required": true,
            "schema": {
              "title": "Review Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReviewItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Candidate Review Item",
        "tags": [
          "review queue"
        ]
      }
    },
    "/v1/candidate-review-items/{review_item_id}/resolution": {
      "post": {
        "description": "checked: you looked at the file, or checked the candidate (every field). dismiss: needs a\nreason. Recorded with your name, once.",
        "operationId": "resolve_candidate_review_item_v1_candidate_review_items__review_item_id__resolution_post",
        "parameters": [
          {
            "in": "path",
            "name": "review_item_id",
            "required": true,
            "schema": {
              "title": "Review Item Id",
              "type": "string"
            }
          },
          {
            "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateResolutionIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateReviewItemOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Resolve Candidate Review Item",
        "tags": [
          "review queue"
        ]
      }
    },
    "/v1/candidates": {
      "get": {
        "description": "Candidates in your scope, newest first, as summaries.\n\n`source` narrows by where a record came from: `?source=cv_upload&source=public_apply` is the\npeople who have arrived since the migration, without the 5,140 that came from the sheet.",
        "operationId": "list_candidates_v1_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "requisition_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 40,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requisition Id"
            }
          },
          {
            "in": "query",
            "name": "owner_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Owner Id"
            }
          },
          {
            "description": "Where the record came from; repeat for more than one",
            "in": "query",
            "name": "source",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "enum": [
                      "tai_master",
                      "cv_upload",
                      "manual_entry",
                      "public_apply"
                    ],
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Where the record came from; repeat for more than one",
              "title": "Source"
            }
          },
          {
            "description": "true: only archived records; false: only the ones still in play",
            "in": "query",
            "name": "archived",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "true: only archived records; false: only the ones still in play",
              "title": "Archived"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatePage"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Candidates",
        "tags": [
          "candidates"
        ]
      },
      "post": {
        "description": "A candidate you found, typed in by hand. Every field is recorded as manual and unchecked,\nwith your name, and the candidate joins the review queue until a person checks them. Fields\nleft out or empty are not recorded. The response carries ids only.",
        "operationId": "enter_candidate_v1_candidates_post",
        "parameters": [
          {
            "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidateEntryIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnteredCandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Enter Candidate",
        "tags": [
          "candidates"
        ]
      }
    },
...
    "/v1/candidates/{candidate_id}": {
      "get": {
        "description": "One candidate, every field with its source and verification (BR-201, BR-202).",
        "operationId": "get_candidate_v1_candidates__candidate_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Candidate",
        "tags": [
          "candidates"
        ]
      }
    },
...
    "/v1/candidates/{candidate_id}/group": {
      "get": {
        "description": "Which record this one is read under, and every record read under it.",
        "operationId": "candidate_group_v1_candidates__candidate_id__group_get",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Candidate Group",
        "tags": [
          "review queue",
          "candidates"
        ]
      }
    },
    "/v1/candidates/{candidate_id}/joins": {
      "post": {
        "description": "Records that two records are one person, with your name and a reason. Both records stay as\nthey are: the joined one is read under this one, and undoing puts it back.",
        "operationId": "join_candidates_v1_candidates__candidate_id__joins_post",
        "parameters": [
          {
            "in": "path",
            "name": "candidate_id",
            "required": true,
            "schema": {
              "title": "Candidate Id",
              "type": "string"
            }
          },
          {
            "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 100,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "A UUID you generate per distinct request. A retry with it returns the first response instead of doing the work again.",
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JoinIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JoinOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "Error: see the code (API plan section 5)"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Join Candidates",
        "tags": [
          "review queue",
          "candidates"
        ]
      }
    },
...
  }
}